requirements.txt
aiodocker==0.21.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
atomicwrites==1.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
check-jsonschema==0.28.5
click==8.1.7
click-default-group==1.2.4
click-didyoumean==0.3.1
croniter==2.0.5
distlib==0.3.8
fasteners==0.19
filelock==3.14.0
flatten-dict==0.4.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
meltano==3.4.2
multidict==6.0.5
numpy==1.26.4
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
prompt-toolkit==3.0.36
psutil==5.9.8
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-ulid==1.1.0
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
referencing==0.35.1
regress==0.4.5
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==69.5.1
six==1.16.0
smart-open==7.0.4
snowplow-tracker==1.0.2
SQLAlchemy==2.0.30
structlog==24.2.0
tabulate==0.9.0
termcolor==2.4.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
virtualenv==20.26.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
yaspin==2.5.0
